Reduce calls to rng with seed=None
By generating a bunch of seeds at the beginning of the module and saving
them in an iterable array, we reduce the amount of calls to the rng with
seed=None, which uses OS entropy. This may help cases where the amount
of entropy is limited (e.g. GitHub Actions containers).